scripts: Use Dpkg::Util list functions instead of grep
This avoids using the grep builtin on scalar context, which is
suboptimal as it does not short-circuit on first match, and it's
intended to produce an output array.

Ideally we'd directly use the functions from List::MoreUtils, because
they might be more performant, but not being a core module is a show
stopper.

Addresses BuiltinFunctions::ProhibitBooleanGrep.