man: Unify references to dpkg project versions for new features
Use the general form of “dpkg VERSION”, w/o any markup, to talk
about the point a feature got introduced in the dpkg project,
regardless of what part of the project was involved, so we use
dpkg in this way as a reference to the entire software project,
and not the dpkg program or the dpkg Debian package.

The name dpkg-dev is a Debian specific packaging layout, which is not
how other downstreams are going to package the software, just avoid
the term. If we need to mention the collection of scripts contained
in the Debian dpkg-dev we use “dpkg build scripts”.

There are still two remaining dpkg-dev references in paths, but those
need to be fixed by substitution at build time, as part of another
commit.