dpkg: Always reset want status when removing or purging a package
Move pkg_set_want() and modstatdb_note() calls from deferred_remove()
to the beginning of the function so that it affects all remove/purge
requests, even if it ends up getting ignored so that we can properly
update the wanted status.

Do not try to set the want status for packages with want_unknown.

Closes: #163763