Dpkg::Arch: Multiple fixes to cope with empty triplets
With the changes introduced in 26a4d1df1119da0a8601b27198086086f9f3249f
debarch_to_debtriplet() and debwildcard_to_debtriplet() can now return an
empty list () instead of (undef) and the logic needs to be adjusted in
multiple place to cope with those changes.

The fixes in this commit might not be exhaustive, they correspond to
warnings that dpkg-source has been emitting. But given the number of
similar changes in the above commit, it's impossible for me to review
everything.