Dpkg::Arch: Properly fix debarch_to_debtriplet() callers
The previous checks have always been shady (the subsequent fix too),
as they didn't guarantee what they were checking for, i.e. that the
array was not missing an element. Now that debarch_to_debtriplet()
returns better values, just check for the correct array size (to be
future-proof), instead of checking for the array not being empty, or
for undefined elements (something that should never happen anymore).