libdpkg: Only ignore older packages if the existing version is informative
When ignoring older packages (either by explicit request, or when
parsing available files), we should not ignore them when the existing
version is empty, because then we would consider that the empty version
(equivalent to "0-0") is newer than "0~0-0" for example.

This affects any program using libdpkg to parse the available file.

Regression introduced in commit 86823d5792af8328107c3d86e486d2cf373da128.

Closes: #676664