dpkg-buildflags: Add GCJFLAGS support
This adds default flags for the GNU Compiler for Java, and the
supported hardened flags. We do not name the variable JFLAGS nor
JAVAFLAGS, to make it very clear these are not flags for javac,
they are only for GCJ; so we use the automake variable name as
existing precedent.

Closes: #708375