dpkg-shlibdeps: New option -l to add private shared library directories
This allows to specify additional build-time shared library directories
without requiring the caller to set LD_LIBRARY_PATH, which is a run-time
dynamic linker variable, and abusing it might be problematic for example
when cross-compiling.

While accepting colon-separated paths might seem tempting, it disallows
valid paths with colons, while not common on system paths, these could
be present on user paths, which makes this a bad interface to have.

Closes: #698881