libdpkg: Fix theoretical stack buffer overflow in w_dependency()
If the dependency field name is longer than 49 chars then there will
be a stack buffer overlow.

But the function is only ever going to be called from known field
names, from the fieldinfos table. So it's currently not possible to
trigger this from the dpkg code base, but fixing it now will avoid
future unpleasant surprises.

Reported-by: Pedro Ribeiro <pedrib@gmail.com>