libdpkg: Reset environment variables for all compressors in the filter
This makes sure the commands will not produce strange output due to
environment settings. The cleaned environment variables are XZ_DEFAULTS,
XZ_OPT, BZIP and BZIP2, depending on the compressor used.

Move GZIP environment variable cleanup to libdpkg, as there's nothing
inherently dpkg-deb specific in resetting the environment variables for
a specific compressor so that the output is reproducible and a bit more
consistent with the code using the specific compressor library.