u-a: When changing slave links do not warn that the link group is broken
When new slave links are added or the targets for existing ones are
changed, we should not give bogus and confusing warnings as if the link
group was broken on disk and needs to be updated. Instead detect these
cases and give an informative message noting that the link group needs
to be updated due to the slave links changes. But still warn in case any
of the slaves is not being updated and is trully broken.