libdpkg: Fallback to use ASCII method for multibyte strings on error
We cannot do strict error checking on multibyte conversion at this
point in time, because supposedly the database is in UTF-8, but the
current locale might be in a different encoding, so the conversion
might fail due to that. It could also fail if the user requested a
standard locale like C or POSIX which does not support handling
UTF-8 encodings.

Regression introduced in commit ab5bdc1fc7bbf225efcbb3376ed309015682e772.

Reported-by: Raphaël Hertzog <hertzog@debian.org>