dpkg: Only keep previous database stat info instead of open files
There's no point in keeping the files open across subsequent calls,
when the only thing we want to know is if the new file is the same as
the old one, just preserve the stat information for that. With this we
do no longer need to set the close-on-exec bit either.