man: Fix wording for ambiguous constructs
Mark in bold and add section to File::Glob reference. Avoid ending a
line with a closing parenthesis as po4a adds two spaces after that.
Use switched instead of converted uniformly when referring to dir to
symlink changes and vice-versa. Refer to the available database as the
available file and mark it in italics to conform to current usage.

Reported-by: Helge Kreutzmann <debian@helgefjell.de>