libdpkg: Fix crashes on Mac OS X when using first gettext() after fork()
On Mac OS X, the libintl code needs to call into the CoreFoundation
framework, which is internally threaded, to initialize some caches.
This is a problem when that first call is done after a fork(3),
because per POSIX, only one thread will survive, leaving the
process in a very inconsistent state, leading to crashes.

To workaround this, we try to force the caches initialization
at program startup time, by performing a dummy gettext("") call.

Tested-by: Martin Costabel <costabel@wanadoo.fr>