dpkg-buildpackage: Avoid unnecessary removal of empty arguments in withecho
The code to remove empty arguments was added to handle a possible
empty rootcommand in commit 669ab06759db350765ac751cb52f487ea3418b7f,
when it was a scalar.

But then rootcommand was switched from a scalar to an array in
commit c2acc4de8e86a9f1ad56c546f16355b796be0c60, which makes the
removal of empty arguments unnecessary as the sublists get automatic
interpolation when calling the function.