Dpkg::Source::Package: Make $diff_ignore_default_regexp a true alias
If the users are modifying this variable, just copying the real one to
the public one will not do it, because changes will not be seen by the
module code. Instead create a real alias.

Regression introduced in commit 5ae4ce9c2dfddbb36e2e0fc24dde3a27196f00f1.