dpkg: Keep the diversions and statoverride database files open
The database files must be kept open during an entire dpkg run, because
we need to check if the files have changed, and rely on the inode number,
which could get eagerly reused by the filesystem if these files get
replaced multiple times in maintainer scripts.

Add a code comment explaining the rationale, so that this does not happen
again in the future.

This partially reverts commit 579b90b61b24a538cf1709d7ceda0b38fec631b3.

Closes: #731524