dpkg-genchanges: Only print build type once
If -B or -A are specified multiple times, there will be multiple
confusing messages stating what will supposedly be including, one
for each option appearence. Just merge the messages with the other
build type description. While at it, reword the binary-only upload
message to be consistent with the new merged ones.