libdpkg: Use makedev(3) rather than ad-hoc computations
The current code does not support large major/minor device numbers, by
using the system's makedev(3) we'll be able to use those. This affects
at least Linux, NetBSD and OpenBSD based systems. In case the function
is not available (such as in Mac OS X), we'll fallback to the compat
implementation with the same current limitations.

Signed-off-by: Peter Chang <dpf@google.com>
Signed-off-by: Guillem Jover <guillem@debian.org>