dpkg-deb: Allow to use the same compression for control.tar as data.tar
Add a new --uniform-compression, that allows to use the same compression
parameters on the control.tar member as for the data.tar member.

This is a transitional need, once a dpkg-deb supporting other control.tar
compressions is widely deployed, ideally on stable distribution releases,
then the default could possibly get switched.