dpkg-deb: Deprecate compressing .deb files with bzip2
This compressor has been superseded by xz when it comes to compression
ratio, and in cases where higher compatibility or raw speed is desired
gzip is still the better option.

Issue a warning for now for the -Zbzip2 option value, the support for
that value will get disabled at a later point.