libdpkg: Switch test suite to use TAP
This has the following benefits. It stops using assert() for the test
framework, it prints a more descriptive message, it performs all test
cases even if one fails. It allows to skip specific cases, or mark as
todo.

This also switches the build system to use the automake TAP driver.