s-s-d: Add a native FreeBSD pid_is_exec() method
Use the KERN_PROC_PATHNAME sysctl interface to retrieve the process
pathname. This will allow to stop requiring the linprocfs fileystem
which is not the native procfs and is not usually mounted by default
anyway.

This still has the problem that the pathname cannot be retrieved when
the inode has been unlinked, in the same way as when accessing the
/proc/<PID>/exe symlink from linprocfs.