build: Inject a Perl coverage index entry into the lcov report
Add correct summary values, create a percentage bar, and remove the
lcov-epilog template, which was being inserted in every and each
generated lcov html file, not just the indices.

The injection should be considered fragile, as it depends on the input
report not changing its structure. But this is no worse than using the
local prolog and epilog html templates.