build: Switch test runner from Test::Harness to TAP::Harness
The latter is the recommended module for new code, the former being just
a compatibility wrapper around it. And it will allow to set more detailed
options for the harness that are not exposed through the Test::Harness
wrapper.

We use a make variable instead of inlining the perl code in the make
recipe due to the way make escapes backslashes inside single-quotes
in recipes, which disturbs the semantics of the perl code.