s-s-d: Fix off-by-one stack buffer overrun on GNU/Linux and GNU/kFreeBSD
Cherry picked from commit 00e2aadcdc9d86655963df13068afd85eca2ed83.

This might happen if the executable pathname is longer than
_POSIX_PATH_MAX. Although this should not have security implications
as the buffer is surrounded by two arrays (so those catch accesses
even if the stack grows up or down), and we are compiling with
-fstack-protector anyway.

We just need to always leave room for the final NUL character.

Warned-by: coverity