dpkg-source: Automatically add the Testsuite field
This is a comma-separated field. The only currently known value for the
field is autopkgtest, which requires the debian/tests/control file to
be present, even if empty, otherwise it is a deb822-style file.

Existing and unknown values will be preserved, and autopkgtest will be
appended. If the autopkgtest value is present but there is no
debian/tests/control file, then the value will be removed and a warning
emitted.