Dpkg::Source::Archive: Delete the current compressor SIGPIPE disposition
With some archives, the tar process will terminate prematurely if it
finds trailing zeros, but the compressor will try to keep feeding it
with data, although the other end of the pipe will now be gone. We
handle this already by ignoring processes terminated by SIGPIPE, but
if the process is ignoring SIGPIPE it might get terminated due to the
error from the write(2) call, which we are not handling. Reset the
SIGPIPE disposition to its default setting.

Closes: #756526