dpkg-genchanges: Fix order of Files field in generated .changes file
Regression introduced in commit e1749d77a44ee1c82c2d047cf9be4b071554832d.

As the code switched to use the list of files to be distributed instead
of the files with checksums, the order changed. Move the code around to
inject the source files first into the distributed files so that we
recover the old behavior, and make it consistent with the Checksums
fields.