dpkg-deb: Stop warning on unknown arbitrary fields
The rest of the tools, particularly the ones in dpkg-dev, do not emit
such warnings, so doing it when building the binary packages is a bit
pointless, and usually annoys maintainers or downstream developers who
end up disabling this code anyway.

This might have served a purpose some time ago, to easily spot typos
in field names, but we do have other means to check for those now,
like more strict parsing in dpkg, which will warn or error on required
fields, or tools like lintian which do a thorough check on the produced
packages.