build: Fix compiler flag detection with non-GCC compilers
We enable -Werror when checking the flags so that on compilers like
clang which just emit a warning on unknown warning options, instead
of erroring out, we get proper detection.

We could just use stuff like -Werror=unknown-warning-option and
-Werror=unused-command-line-argument, but then we would need to detect
if those are available as they might make the compiler error out if
they are no supported. So just using the blanket -Werror, although more
broad, makes the code less complicated.