s-s-d: On kFreeBSD do not make kvm_openfiles() open /dev/mem
We do not need that file, as we are not using any kvm_read(3) call,
either explicitly or implicitly. Everything is retrieved through
sysctl(2).

libkvm from FreeBSD allows to pass /dev/null as the memory file to
denote that we do not need it.

This will make s-s-d work again on jails which do not usually have
/dev/mem available.

Reported-by: Steven Chamberlain <steven@pyro.eu.org>