dpkg-shlibdeps: Initialize with correct minimum version from a symbols file
When initializing the minimum version for a dependency, we have to use
the greater one from any SONAME present in that symbols file, because
the shared libraries might not have been available before that version
for example.

This also guarantees a reproducible output, as the minimum version
selected is not dependant on the hash order.

Closes: #764721

[guillem@debian.org:
 - Reword debugging message. ]

Signed-off-by: Guillem Jover <guillem@debian.org>