s-s-d: Create a pidfile even when not backgrounding ourselves
Regression introduced in commit 29778da537e2ff1a0f032db33dde43413b7345ef.

When the user asked us to create the pidfile, but _not_ to background
ourselves, we should still create the pidfile.

This usage is somewhat dubious, as s-s-d has an option to background,
which is better in any way, as it will make sure to return error codes
in case the program cannot be started for whatever reason. But it's
still a regression.

Closes: #765110