dpkg: Add explicit await trigger directives and command-line options
The current default of awaiting is really bad, because most packages
do not really require to be put in a triggers-awaited status. Also
maintainers tend to use it instead of the -noawait variant, making
trigger cycles more common.

Changing the default behavior was not an option, as precisely doing
so would affect the packages that would break when not put in a
triggers-awaited status.

Instead we'll add explicit await variants for the triggers, and
dpkg-trigger, and probably deprecate the implicit ones during the
1.18.x cycle. Maybe, and only maybe it can be switched to be the
noawait variants in 1.19.x or 1.20.x, perhaps.