dpkg: Make sure to always switch to the next dependtry after no progress
We might get into a situation where we are making no progress, and
dependencies_ok() is still telling us that we can make progress by
fixing some triggers awaited package, but if those packages require
a dependency cycle to be broken we will get stuck and not progress to
the next dependtry.

Add an additional hard limit where we will jump dependtry, in case we
are not making any progress. This is a safety net that was removed when
triggers got introduced, but that only became active when adding the
dependency checks to trigproc(), although it was never future-proof.

Regression introduced in commit 10440009b68f59eeed4cb1b56547e3cf356aa540
and commit e121d882c92099d1f1aa9eedd0c0e90806a44efa.

Closes: #766242, #766322