debian: Add versioned Breaks on packages creating trigger cycles
There are two groups of packages, ones where the old version of the
packages produced trigger cycles as they were using awaiting trigger
directives, but are fixed now, which get a << versioned Breaks on the
fixed version. And packages that are still not fixed that get a <=
versioned Breaks on the broken version.