libdpkg: Change m_calloc() signature to take also a nmemb argument
We are using these on arrays, so we end up having to compute the total
size from the number of members and the member size. Just delegate the
computation to the underlaying call which should ideally also handle
overflows and similar. And we can always pass nmemb=1 from the call
sites if needed.