dpkg: Pass <new-version> to maintscript actions that cannot get it otherwise
The upgrade actions and their rollback counterparts do not have access
to the new binary package which is somewhere on the filesystem depending
on how dpkg was invoked, nor to its metadata. If the maintainer script
wants to know the <new-version>, it needs to be turned into a template
and that version instantiated at build-time, which is cumbersome.

So we now pass <new-version> when these upgrade actions are already
getting <old-version>, but not on the cases where they do not get any
argument at all, as that's their differentiator.

Prompted-by: Andrey Utkin <andrey.krieger.utkin@gmail.com>