dpkg: Use utimensat() if available and in preference of lutimes/utimes
This funcion is only specified in POSIX.1-2008, but if it is, it is
going to be more standard than lutimes() which is only available on
Linux and the BSDs.

XXX: On recent Linux (2.6.32, 2009-12) and Solaris (11.1, 2013-04)
systems utimensat() is pretty much broken on multiple conditions.
→ But these issues (taken from gnulib and man utimensat(), do not really
affect our usage, no UTIME_OMIT nor UTIME_NOW, etc.