Dpkg::Dist::Files: Allow ~ in added filenames
When we are constructing a .changes file we add the source artifacts
to the distribution manually, before loading the distribution list
from debian/files.

We should allow ~ in the filename as that's a valid character in a
version number.

Regression introduced in commit ab15fd24dd1a8207ab1463410338f24283989f7c.

Reported-by: Niels Thykier <niels@thykier.net>