Dpkg::Shlibs: Do not add cross-root directories to default search list
There's no reason to include these in the default search library list
anymore, as multiarch has superseded this usage for a long time, and no
package in Debian should be shipping libraries in those paths neither
other packages linking against those. Not including the paths avoids
polluting the search list.

If a user outside Debian wants to use a cross-root layout, the
«dpkg-shlibdeps -l» option can be used to specify those.