libdpkg: Consider foreign packages ambiguous in need of arch-qualifier
With cross-arch dependencies, foreign arch-qualified dependencies and
foreign packages become really ambiguous in error messages, but also
on the usual progress reporting.

Arch-qualifying foreign packages should be backwards compatible, because
if a user had foreign packages installed on a pre-multiarch dpkg, then
that was already out of spec. And if they do now, it means they have
enabled multiarch. Keeping Multi-Arch:same packages always arch-qualified
still makes sense because those will not appear on a non-Multi-Arch
enabled distribution, and are required to distinguish which instance we
are referring to.