dpkg-checkbuilddeps: Add support for ignoring built-in build dependencies
There is currently no way to require that «Build-Essential: yes»
packages be installed, so on Debian-based systems we have to rely on
the build-essential metapackage, which is slightly wrong.

In addition users that know what they are doing, might want to disable
only the implicit build dependencies and conflicts, but not the entire
dependency checks, and there was no way to do that.

Add new «dpkg-buildpackage --ignore-builtin-builddeps» and
«dpkg-checkbuilddeps -I» options.

Closes: #480638, #571671