dpkg-deb: Set ar timestamp and tar mtime clamping to SOURCE_DATE_EPOCH
The ar and tar file entries contain a timestamp which is usually filled
with the current build time and the file mtime respectively. This causes
the generated deb(5) packages to not be reproducible for the same source
package and same build environment.

We now check if the environment variable SOURCE_DATE_EPOCH is defined,
parse that and use it as the timestamp for ar and to clamp the mtime in
the tar file entries.

This variable is specified to contain the timestamp of the source
release date in seconds since the epoch.

Closes: #759999
Base-on-patch-by: Jérémy Bobbio <lunar@debian.org>