scripts/mk: Add an extra level of escaping for double evals
Regression introduced in commit 486241a9d3521ed093be9a02b00f4d404dc39b35.

Otherwise the values are not computed lazily, many of which are quite
expensive. Rename dpkg_late_eval to dpkg_lazy_eval so that the other
makefiles can be intermixed with the ones needing the extra level of
escaping.

Closes: #793330