debian: Use --with-devlibdir to only switch libdpkg-dev to a multiarch dir
The perl plugins and dselect methods need to use the non-multiarch
directory, or things break. We can use now the new option to only
install the C libdpkg files in a multiarch enabled path.

The fix from commit 3ce3a7adeb8c815734a35b95aee96bacd1aafde4 was not
correct, as libdpkg.pc was not moved to the multiarch directory and
was making the build fail, because it needs to have a libdir matching
the directory where the library was installed.