u-a: Merge all action handling into a single if-else-if block
Introduce two boolean variables to track when to log the command-line
arguments when modifying anything on the system, and another one when
modifying a single alternative to initialize and update it.

This unifies the code, and will make it easier to switch it into a
shared library in the future.