dpkg: Export variable DPKG_ROOT in maintainer scripts
This variable holds the value of instdir. It is supposed to be used
in maintainer scripts. It should be prepended to all paths that are
operated on. Currently, dpkg chroots to the instdir before invoking
maintainer scripts, so when it does that DPKG_ROOT is set to the empty
string. Thus currently, DPKG_ROOT is always empty.

[guillem@debian.org:
 - Reword man page to clarify current status. ]

Closes: #804624
Signed-off-by: Guillem Jover <guillem@debian.org>