Dpkg::Source: Make patching a file multiple times fatal for first quilt patch
When we check if the first quilt patch in a series can be applied, we
use «patch --dry-run» with other options, but that cannot work if the
same diff patches a file multiple times, as patch does not record the
previous state in memory.

We are alredy detecting this condition, but only emitting a warning,
instead make it possible to turn it into a fatal error.

Closes: #810720
Reported-by: Apollon Oikonomopoulos <apoikos@debian.org>