Add MIPS R6 architectures to arch tables
MIPS R6 is a new release of MIPS32 and MIPS64. R6 is not fully compatible
with R5-, as it adds and *removes* some instructions, and adds emulation
of the removed instructions in the kernel, so old binaries can still run
on new R6 CPUs, but at a significant performance cost.

Closes: #807340
Signed-off-by: Guillem Jover <guillem@debian.org>