u-a: Remove useless alternative_has_slave() before alternative_get_slave()
This gets rid of a warning with gcc-6, that does not understand that we
never pass a NULL ‘sl’ pointer as we have checked previously if it has
slaves via alternative_has_slave().

And gets rid of unneeded calls to alternative_has_slave(), which
internally do alternative_get_slave() anyway.