u-a: Generalize admindir environment variable handling
This variable is currently set by dpkg, but other package managers or
callers might set it with a different name on other downstream projects.
Let's move the variable name to the build system so that third parties
do not have to modify the code going on to remove a dpkg-specific name.