dpkg-shlibdeps: Add new -I option to ignore package build directories
This option can be used to completely ignore local package build trees,
for example in cases where've got multiple compatible builds of the same
shared library, which provide use the same SONAME, and we want all our
local objects to use the same single library instead of whatever came
first. Using -x in the above case does not work, because that just
removes the dependency after processing, so we end up with missing
dependencies.

Closes: #821025