dpkg: Fix file queue tail assignment in file queue pop
When we are popping a file from the files queue, we might end up messing
it up. To check for the side-effects on the file-system one should look
for files with .dpkg-new extension for packages that have been fully
configured and where those files are not listed as owned by them. These
packages will need to be reinstalled.

Closes: #823288