dpkg-source: Add new --require-strong-checksums option and change default
Erroring out when no strong checksums are present is very harsh, as we
do not even do something similar for invalid/unknown/expired signatures
which means doing this for checksums has really no point.

Add a new command-line option to force the behavior to be strict, and
change to a warning.

Regression introduced in commit 040973c7a1e50b78ef042ef5ffbfff0440c24700.

Closes: #823428
Reported-by: Niko Tyni <ntyni@debian.org>