dpkg-buildpackage: Preset build timestamp to latest changelog entry
Set the SOURCE_DATE_EPOCH environment variable to Unix timestamp since
the epoch of the latest entry in debian/changelog. This enables build
to be easily reproduced as the dates captured in the «.deb» archives
will be deterministic.

SOURCE_DATE_EPOCH can also be externaly set before running
dpkg-buildpackage to reproduce the build with an arbitrary date.

Closes: #759999
Ref: https://reproducible-builds.org/specs/source-date-epoch/
Base-on-patch-by: Jérémy Bobbio <lunar@debian.org>