scripts: Replace changelog program parsers with perl modules
Using programs to implement the custom changelog parsers was very
inefficient as it required to parse the custom changelog, output deb822
formatted entries to then parse that and output again with the desired
format.

These were implemented as programs because at the time the perl code
in dpkg was not using perl modules, so it was not easy to extend. Using
perl modules now is cleaner and allows for a faster implementation.

In addition there's no known users in Debian, so it was deemed safe to
remove the support without a transition.